cloud_run/fetch_realtime/requirements.txt
functions-framework==3.*
flask==3.0.*
gunicorn==21.*
websocket-client==1.7.*
google-cloud-storage>=2.14.0,<3.0.0
google-cloud-pubsub>=2.25.0,<3.0.0